+++ ReleaseNotes.md
@@ -44,3 +44,4 @@
 - Add Mile::ScopeExitTaskHandler template class.
 - Add Mile::SplitCommandLineWideString function.
 - Add Mile::SplitCommandLineString function.
+- Add MileEnumerateFileByHandle function.